Renaming and Reorganizing Networking module
