+++ src/Messaging/Commands/SendPassengersCommand.java
@@ -1,7 +1,7 @@
-package Networking.Messages;
+package Messaging.Commands;
 
-import Networking.Direction;
-import Networking.Transmitters.DMA_Transmitter;
+import Messaging.Direction;
+import Messaging.Transmitters.DMA_Transmitter;
 
 /**
  * SendPassengersCommand Record, Command to floor to send passengers back through the provided transmitter.
