+++ test/unit/Networking/DMA_ReceiverTest.java
@@ -1,9 +1,10 @@
 package unit.Networking;
 
-import Networking.Direction;
-import Networking.Messages.DestinationEvent;
-import Networking.Messages.SystemEvent;
-import Networking.Receivers.DMA_Receiver;
+import Messaging.Direction;
+import Messaging.Events.DestinationEvent;
+import Messaging.Events.SystemEvent;
+import Messaging.Receivers.DMA_Receiver;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -12,13 +13,12 @@ class DMA_ReceiverTest {
 
     DMA_Receiver receiver;
     SystemEvent systemEvent;
-    @Test
+    @BeforeEach
     void setUp() {
         this.receiver = new DMA_Receiver();
     }
     @Test
     void setAndReceiveMessage() {
-        setUp();
         systemEvent = new DestinationEvent(2, Direction.UP);
         receiver.setMessage(systemEvent);
         DestinationEvent event = (DestinationEvent) receiver.receive();
