+++ test/unit/Networking/README.md
@@ -2,7 +2,7 @@
 * Last Edited: 2024/02/03
 
 ## Description
-Unit tests for applicable Networking classes.
+Unit tests for applicable Messaging classes.
 
 ## Contents
 * `DMA_ReceiverTest.java`
